requirements.txt
cached_property
click
imageio
jupyter
nbconvert
notebook

# Maths
numpy

# Image Processing.
Pillow

# TODO: Remove, these are really just for development.
# autopep8
# black
# isort
# pycodestyle
# pydocstyle

pytest
scikit-image

##
# Host Environment Requirements
##
#16.04 18.04: git
#16.04 18.04: python3
#16.04 18.04: python3-venv
#16.04 18.04: python3-dev

#git+https://github.com/jed-frey/jupyter_NotebookCleaners.git#egg=notebookcleaner
